Code style fix
