+++ src/main/java/ru/yandex/practicum/filmorate/dao/InMemoryFilmStorage.java
@@ -14,7 +14,6 @@ import ru.yandex.practicum.filmorate.model.Mpa;
 import java.time.LocalDate;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
 
 @Component
